Reformatted pencil marks in light mode to be more readable
This change makes the pencil marks made in the center a dark purple
while also making the font weight bold. The outer ring pencil marks are
changed to gray although the gray might be a little too light.

The goal of this is to make the center pencil marks stand out from the
rest of the marks. On more difficult puzzles this helps to input other
possible values into the cells and have some differentiation between the
center marks and the outer ring marks.

To make the numbers readable when the cell is selected, the cell colour
was also changed to a light yellow. Again, possibly too low contrast but
it works for now.